60184 - invalid OOXML produced when XSSFFont.getFamily() called with no families defined
Fix to stop creating an empty family when none exist, just use the existing logic to return NOT_APPLICABLE.  Creating the empty element produces XML that doesn't comply with the XSD.

Alternative would have been to set the family int value for the new family, but this way checking the value doesn't change the output, which is my preferred behavior.  No quantum effects!

git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1794111 13f79535-47bb-0310-9956-ffa450edef68